+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Please, check out guidelines: https://keepachangelog.com/en/1.0.0/
 
 ## Next
 
+- Fix failed `tuist dependencies fetch/update`  command when `Carthage` dependency is imported as binary [#3108](https://github.com/tuist/tuist/pull/3164) by [@havebenfitz](https://github.com/havebeenfitz)
+
 ## 1.45.1
 
 ### Fixed